Renaming module from multitenancy to context propagation
