+++ stream-dbs-clients/src/main/java/com/backbase/stream/clients/autoconfigure/DbsApiClientsAutoConfiguration.java
@@ -12,7 +12,7 @@ import com.backbase.stream.clients.config.PortfolioApiConfiguration;
 import com.backbase.stream.clients.config.TransactionManagerClientConfig;
 import com.backbase.stream.clients.config.UserManagerClientConfig;
 import com.backbase.stream.clients.config.UserProfileManagerClientConfig;
-import com.backbase.stream.m10y.config.MultiTenancyConfiguration;
+import com.backbase.stream.context.config.ContextPropagationConfiguration;
 import com.backbase.stream.webclient.configuration.DbsWebClientConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
@@ -28,7 +28,7 @@ import org.springframework.context.annotation.Import;
 @Configuration
 @Import({
     DbsWebClientConfiguration.class,
-    MultiTenancyConfiguration.class,
+    ContextPropagationConfiguration.class,
     AccessControlClientConfig.class,
     ApprovalClientConfig.class,
     ArrangementManagerClientConfig.class,
