+++ stream-sdk/stream-parent/stream-context-propagation/src/main/java/com/backbase/stream/context/TenantContext.java
@@ -1,4 +1,4 @@
-package com.backbase.stream.m10y;
+package com.backbase.stream.context;
 
 import java.util.Optional;
 import lombok.extern.slf4j.Slf4j;
